Use URI for "createController" command (#397)
According to the LSP specification, the "CreateFile"
operation uses the `DocumentUri` type for the resource name:

```json
{
  "kind": "create",
  "uri": "file:///Users/file.js"
}
```

However, the "createController" command currently uses
an absolute path, which breaks compatibility with
clients other than VSCode. To comply with the specification
and make this feature working in other clients,
the command must use a URI in the "CreateFile" operation.